+++ gulpfile.js
@@ -8,6 +8,7 @@ const browserSync = require('browser-sync').create();
 const clean = require('gulp-clean');
 const concat = require('gulp-concat');
 const uglify = require('gulp-uglify');
+const ts = require("gulp-typescript");
 
 const SRC = "src";
 
@@ -16,6 +17,7 @@ const PATHS = {
     dist: 'dist',
     scss: `${SRC}/scss/**/*.scss`,
     js: `${SRC}/scripts/**/*.js`,
+    ts: `${SRC}/scripts/**/*.ts`,
     html: `${SRC}/**/*.html`,
     images: `${SRC}/assets/images/**/*.*`
 };
@@ -43,6 +45,7 @@ function serve() {
     watch('src/scss/**/*.scss', buildScss);
     watch('src/**/*.html', buildHtml);
     watch(PATHS.js, buildJs);
+    watch(PATHS.ts, buildTs);
 }
 
 function createDevServer() {
@@ -80,10 +83,30 @@ function buildJs() {
         .pipe(browserSync.stream());
 }
 
+function buildTs() {
+    return src(PATHS.ts)
+      .pipe(
+        ts({
+          target: "ES5",
+          module: "CommonJS",
+        })
+      )
+      .pipe(dest(`${PATHS.src}/js`))
+      .pipe(dest(`${PATHS.dist}/js`))
+      .pipe(browserSync.stream());
+  }
+
 exports.sass = buildScss;
 exports.html = buildHtml;
 exports.copy = copy;
 exports.cleanDist = cleanDist;
 
-exports.build = series(cleanDist, buildScss, buildJs, buildHtml, copy);
-exports.default = series(buildScss, buildJs, parallel(createDevServer, serve));
+exports.build = series(
+    cleanDist,
+    parallel(buildScss, buildJs, buildTs, buildHtml, copy)
+  );
+  exports.default = series(
+    parallel(buildScss, buildJs, buildTs, buildHtml, copy),
+    parallel(createDevServer, serve)
+  );
+
