+++ src/pages/army_DA.html
@@ -90,6 +90,7 @@
     
     <script src="https://cdn.jsdelivr.net/npm/gridjs/dist/gridjs.umd.js"></script>
     <script src="../js/bundle.js"></script>
+    <script src="../js/typescript.js"></script>
 </body>
 
 </html>
