+++ src/scripts/index.js
@@ -1,24 +1,24 @@
 console.log("Hello world!")
 
-var imgs = document.querySelectorAll(".image__toggled");
-    imgs.forEach((img) => {
-            img.addEventListener("click", toggleImage);
-        }
-    )
-function toggleImage(event) {
-    var img =event.target;
-        img.classList.toggle("image_toggled__on");
-    }
-
-
-    //gridjs check
-    new gridjs.Grid({
-        columns: ["Name", "Email", "Phone Number"],
-        data: [
-          ["John", "john@example.com", "(353) 01 222 3333"],
-          ["Mark", "mark@gmail.com", "(01) 22 888 4444"],
-          ["Eoin", "eoin@gmail.com", "0097 22 654 00033"],
-          ["Sarah", "sarahcdd@gmail.com", "+322 876 1233"],
-          ["Afshin", "afshin@mail.com", "(353) 22 87 8356"]
-        ]
-      }).render(document.getElementById("table"));
+//var imgs = document.querySelectorAll(".image__toggled");
+//    imgs.forEach((img) => {
+//            img.addEventListener("click", toggleImage);
+//        }
+//    )
+//function toggleImage(event) {
+//    var img =event.target;
+//        img.classList.toggle("image_toggled__on");
+//    }
+//
+//
+//    //gridjs check
+//    new gridjs.Grid({
+//        columns: ["Name", "Email", "Phone Number"],
+//        data: [
+//          ["John", "john@example.com", "(353) 01 222 3333"],
+//          ["Mark", "mark@gmail.com", "(01) 22 888 4444"],
+//          ["Eoin", "eoin@gmail.com", "0097 22 654 00033"],
+//          ["Sarah", "sarahcdd@gmail.com", "+322 876 1233"],
+//          ["Afshin", "afshin@mail.com", "(353) 22 87 8356"]
+//        ]
+//      }).render(document.getElementById("table"));
